+++ materials/test_attachment_selectors.py
@@ -33,9 +33,11 @@ class PersonAttachmentSelectorApiTests(SimpleTestCase):
             selectors.__all__,
             (
                 "get_event_attachment_links",
+                "get_event_source_links",
                 "get_person_attachment_links",
                 "get_person_name_source_links",
                 "get_visible_event_attachment_links",
+                "get_visible_event_source_links",
                 "get_visible_person_attachment_links",
                 "get_visible_person_name_source_links",
             ),